Fixes and reorg
