+++ Cargo.toml
@@ -11,13 +11,14 @@ members = [
 	"parity-bytes",
 	"parity-crypto",
 	"parity-path",
-	"patricia_trie",
+	"patricia-trie",
 	"plain_hasher",
 	"rlp",
+	"substrate-trie",
 	"transaction-pool",
 	"trace-time",
 	"triehash",
 	"trie-standardmap",
-	"triestream",
-	"uint"
+	"uint",
+	"test-support/patricia-trie-ethereum"
 ]
