+++ patricia-trie/Cargo.toml
@@ -12,7 +12,6 @@ log = "0.3"
 rand = "0.4"
 hashdb = { version = "0.3", path = "../hashdb" }
 parity-bytes = { version = "0.1", path = "../parity-bytes" }
-triestream = { path = "../triestream", features = ["codec"] }
 parity-codec = "~2.0.1"
 
 [dev-dependencies]
@@ -24,7 +23,6 @@ rlp = { version = "0.3.0", path = "../rlp", default-features = false }
 trie-standardmap = { version = "0.1", path = "../trie-standardmap", default-features = false }
 triehash = { version = "0.4", path = "../triehash", default-features = false }
 parity-bytes = { version = "0.1.0", path = "../parity-bytes" }
-triestream = { version = "0.1.0", path = "../triestream", features = ["ethereum"]}
 # REVIEW: what's a better way to deal with this? The tests here in
 # `patricia_trie` use `keccak-hasher` and `patricia-trie-ethereum` to
 # instantiate concrete impls. Neither crate is needed/wanted in `parity-common`
