+++ patricia-trie/benches/trie.rs
