+++ patricia-trie/src/lib.rs
@@ -20,7 +20,6 @@ extern crate parity_bytes as bytes; // TODO: name changed; update upstream when 
 extern crate hashdb;
 extern crate rand;
 extern crate parity_codec as codec;
-extern crate triestream;
 #[macro_use]
 extern crate log;
 
@@ -40,11 +39,8 @@ extern crate rlp;
 extern crate keccak_hash as keccak;
 #[cfg(test)]
 extern crate keccak_hasher;
-#[cfg(test)]
-extern crate triehash;
 
 use std::{fmt, error};
-use hashdb::{HashDB, Hasher};
 use std::marker::PhantomData;
 
 pub mod node;
@@ -60,9 +56,8 @@ mod lookup;
 mod nibblevec;
 mod nibbleslice;
 mod node_codec;
-mod codec_error;
-mod parity_node_codec;
 
+pub use hashdb::{HashDB, Hasher};
 pub use self::triedb::{TrieDB, TrieDBIterator};
 pub use self::triedbmut::{TrieDBMut, ChildReference};
 pub use self::sectriedbmut::SecTrieDBMut;
