+++ patricia-trie/src/lookup.rs
