+++ patricia-trie/src/nibblevec.rs
