+++ patricia-trie/src/node_codec.rs
