+++ patricia-trie/src/triedbmut.rs
@@ -975,7 +975,7 @@ mod tests {
 	use memorydb::MemoryDB;
 	use rlp::{Decodable, Encodable};
 	use triehash::trie_root;
-	use triestream::RlpTrieStream;
+	use patricia_trie_ethereum::RlpTrieStream;
 	use standardmap::*;
 	use ethtrie::{TrieDBMut, RlpCodec, trie::{TrieMut, NodeCodec}};
 	use env_logger;
