+++ substrate-trie/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "substrate-trie"
+version = "0.4.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Patricia trie stuff using a parity-codec node format"
+repository = "https://github.com/paritytech/parity-common"
+license = "GPL-3.0"
+
+[dependencies]
+parity-codec = { version = "~2.0.1" }
+patricia-trie = { path = "../patricia-trie" }
+triehash = { path = "../triehash" }
+
+[dev-dependencies]
+keccak-hasher = { version = "0.1", path = "../test-support/keccak-hasher" }
+memorydb = { path = "../memorydb" }
