+++ substrate-trie/README.md
@@ -1,0 +1,2 @@
+This crate provides utility functions to validate and initialize tries using flexible input.
+It is used extensively in `parity-ethereum` to validate blocks (mostly transactions and receipt roots).
