+++ test-support/patricia-trie-ethereum/Cargo.toml
@@ -6,7 +6,7 @@ description = "Merkle-Patricia Trie (Ethereum Style)"
 license = "GPL-3.0"
 
 [dependencies]
-patricia-trie = { version = "0.3", path = "../../patricia_trie" }
+patricia-trie = { version = "0.3", path = "../../patricia-trie" }
 keccak-hasher = { version = "0.1", path = "../keccak-hasher" }
 hashdb = { version = "0.3", path = "../../hashdb" }
 rlp = { version = "0.3.0", path = "../../rlp" }
