+++ test-support/patricia-trie-ethereum/src/lib.rs
@@ -60,3 +60,121 @@ pub type TrieFactory = trie::TrieFactory<KeccakHasher, RlpCodec>;
 pub type TrieError = trie::TrieError<H256, DecoderError>;
 /// Convenience type alias for Keccak/Rlp flavoured trie results
 pub type Result<T> = trie::Result<T, H256, DecoderError>;
+
+#[cfg(test)]
+mod tests {
+	use super::{trie_root, sec_trie_root, shared_prefix_len};
+	use super::unhashed_trie;
+	use keccak_hasher::KeccakHasher;
+	use patricia_trie_ethereum::RlpTrieStream;
+	
+	#[test]
+	fn sec_trie_root_works() {
+		let v = vec![
+			("doe", "reindeer"),
+			("dog", "puppy"),
+			("dogglesworth", "cat"),
+		];
+		assert_eq!(
+			sec_trie_root::<KeccakHasher, RlpTrieStream, _, _, _>(v.clone()),
+			"d4cd937e4a4368d7931a9cf51686b7e10abb3dce38a39000fd7902a092b64585".into(),
+		);
+	}
+
+	#[test]
+	fn trie_root_works() {
+		let v = vec![
+			("doe", "reindeer"),
+			("dog", "puppy"),
+			("dogglesworth", "cat"),
+		];
+		assert_eq!(
+			trie_root::<KeccakHasher, RlpTrieStream, _, _, _>(v),
+			"8aad789dff2f538bca5d8ea56e8abe10f4c7ba3a5dea95fea4cd6e7c3a1168d3".into()
+		);
+		assert_eq!(
+			trie_root::<KeccakHasher, RlpTrieStream, _, _, _>(vec![
+				(b"A", b"aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa" as &[u8])
+			]),
+			"d23786fb4a010da3ce639d66d5e904a11dbc02746d1ce25029e53290cabf28ab".into()
+		);
+	}
+
+	#[test]
+	fn test_triehash_out_of_order() {
+		assert!(trie_root::<KeccakHasher, RlpTrieStream, _, _, _>(vec![
+			(vec![0x01u8, 0x23], vec![0x01u8, 0x23]),
+			(vec![0x81u8, 0x23], vec![0x81u8, 0x23]),
+			(vec![0xf1u8, 0x23], vec![0xf1u8, 0x23]),
+		]) ==
+		trie_root::<KeccakHasher, RlpTrieStream, _, _, _>(vec![
+			(vec![0x01u8, 0x23], vec![0x01u8, 0x23]),
+			(vec![0xf1u8, 0x23], vec![0xf1u8, 0x23]), // last two tuples are swapped
+			(vec![0x81u8, 0x23], vec![0x81u8, 0x23]),
+		]));
+	}
+
+	#[test]
+	fn test_shared_prefix() {
+		let a = vec![1,2,3,4,5,6];
+		let b = vec![4,2,3,4,5,6];
+		assert_eq!(shared_prefix_len(&a, &b), 0);
+	}
+
+	#[test]
+	fn test_shared_prefix2() {
+		let a = vec![1,2,3,3,5];
+		let b = vec![1,2,3];
+		assert_eq!(shared_prefix_len(&a, &b), 3);
+	}
+
+	#[test]
+	fn test_shared_prefix3() {
+		let a = vec![1,2,3,4,5,6];
+		let b = vec![1,2,3,4,5,6];
+		assert_eq!(shared_prefix_len(&a, &b), 6);
+	}
+
+	#[test]
+	fn learn_rlp_trie_empty() {
+		let input: Vec<(&[u8], &[u8])> = vec![];
+		let trie = unhashed_trie::<KeccakHasher, RlpTrieStream, _, _, _>(input);
+		println!("[learn_rlp_trie_empty] 1st byte of trie:\n{:#010b}\n trie: {:#x?}", trie[0], trie );
+		assert_eq!(trie, vec![0x80]);
+	}
+
+	#[test]
+	fn learn_rlp_trie_single_item() {
+		let input: Vec<(&[u8], &[u8])> = vec![(&[0x13], &[0x14])];
+		let trie = unhashed_trie::<KeccakHasher, RlpTrieStream, _, _, _>(input);
+		println!("[learn_rlp_trie_single_item] 1st byte of trie:\n{:#010b}\n trie: {:#x?}", trie[0], trie );
+		assert_eq!(trie, vec![0xc4, 0x82, 0x20, 0x13, 0x14]);
+		// The key, 0x13, as nibbles: [ 0x1, 0x3 ]
+		// build_trie will call append_leaf with k/v: [ 0x1, 0x3 ], [0x14]
+		// 	append_leaf will call rlp begin_list(2)
+		// 		begin_list adds 0 to buffer - modified later when list is closed
+		//	key is hpe'd: even length, leaf (terminated) => high nibble sets termination bit, low nibble is zero => 0b0010_0000 => 0x20 => 32
+		// 	append_iter() is called with hpe byte + key byte => 0x20, 0x13; adds 0x80 + length of items (2) => 0x82
+		//	buffer is now: 0, 0x82, 0x20, 0x13, 0x14
+		//	append() adds the value bytes => 0x14 and closes the list: 0xc0 + length of payload => 0xc0 + 4
+		// final buffer: 0xc4 0x82 0x20 0x13 0x14
+	}
+
+	#[test]
+	fn learn_rlp_trie_single_item2() {
+		let input = vec![(
+			vec![0x12, 0x12, 0x12, 0x12, 0x13, 0x13], 	// key
+			vec![0xff, 0xfe, 0xfd, 0xfc]				// val
+		)];
+		let trie = unhashed_trie::<KeccakHasher, RlpTrieStream, _, _, _>(input);
+		// println!("[learn_rlp_trie_single_item] 1st byte of trie:\n{:#010b}\n trie: {:#x?}", trie[0], trie );
+		assert_eq!(trie, vec![
+			0xc0 + 13,	// list marker + 13 bytes long payload
+			0x80 + 7,	// value marker + 7 bytes long payload
+			0x20, 		// HPE byte
+			0x12, 0x12, 0x12, 0x12, 0x13, 0x13,
+			0x80 + 4, 	// value marker + 4 bytes long payload
+			0xff, 0xfe, 0xfd, 0xfc
+		]);
+	}
+}
