--- triestream/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "triestream"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-description = "Trait used to build a byte-stream to calculate trie roots"
-repository = "https://github.com/paritytech/parity-common"
-license = "GPL-3.0"
-
-[dependencies]
-hashdb = { version = "0.3", path = "../hashdb" }
-hex-prefix-encoding = { version = "0.1", path = "../hex-prefix-encoding" }
-rlp = { version = "0.3", path = "../rlp", optional = true, default-features = false }
-parity-codec = { version = "~2.0.1", optional = true }
-
-[features]
-default = ["codec"]
-ethereum = ["rlp"]
-codec = ["parity-codec"]
